# docs/openapi.yaml
openapi: 3.1.0
info:
  title: PayMux API
  version: "1.0.0"
  summary: Open-source payment gateway multiplexer and webhook router.
  description: |
    PayMux lets several applications share one payment gateway account. Each
    application authenticates with its own API key, which is what determines
    ownership: PayMux never trusts an application identifier supplied in a
    request body.

    **Amounts** are integers in the currency's minor unit. For IDR that is one
    rupiah, so `150000` means Rp 150.000.

    **Identifiers** are prefixed, sortable ULIDs: `pay_…` for payments,
    `evt_…` for events, `dlv_…` for deliveries.
  license:
    name: Apache-2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
  contact:
    name: PayMux
    url: https://github.com/anggapixa/paymux
servers:
  - url: https://paymux.example.com
    description: Your PayMux deployment
  - url: http://localhost:8080
    description: Local development
security:
  - ApiKeyAuth: []
...
paths:
  /api/v1/payments:
    post:
      tags: [Payments]
      summary: Create a payment
      description: |
        Opens a payment at the gateway and returns the checkout token and
        redirect URL. Send an `Idempotency-Key` so a retried request returns
        the original payment instead of opening a second transaction.
      operationId: createPayment
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePaymentRequest"
            examples:
              snap:
                summary: A Snap payment
                value:
                  application_order_id: INV-000123
                  amount: 150000
                  currency: IDR
                  customer:
                    first_name: John
                    last_name: Doe
                    email: john@example.com
                    phone: "+628123456789"
                  items:
                    - id: PROD-001
                      name: Example Product
                      price: 150000
                      quantity: 1
      responses:
        "201":
          description: The payment was created.
          headers:
            Idempotent-Replay:
              description: Present and `true` when this is a replayed response.
              schema: { type: string }
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payment"
        "401": { $ref: "#/components/responses/Unauthorized" }
        "409": { $ref: "#/components/responses/Conflict" }
        "422": { $ref: "#/components/responses/ValidationFailed" }
        "503": { $ref: "#/components/responses/GatewayNotConfigured" }

    get:
      tags: [Payments]
      summary: List payments
      description: Returns the calling application's payments, newest first.
      operationId: listPayments
      parameters:
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/StartingAfter"
        - name: status
          in: query
          schema: { $ref: "#/components/schemas/PaymentStatus" }
        - name: application_order_id
          in: query
          schema: { type: string }
        - name: created_from
          in: query
          schema: { type: string, format: date-time }
        - name: created_to
          in: query
          schema: { type: string, format: date-time }
      responses:
        "200":
          description: A page of payments.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentList"
        "401": { $ref: "#/components/responses/Unauthorized" }
...
  /api/v1/payments/{paymentId}/snap/cancel:
    parameters:
      - $ref: "#/components/parameters/PaymentId"
    post:
      tags: [Payments]
      summary: Cancel the checkout session
      description: |
        Retires the Snap checkout session so its token can no longer be paid.
        Midtrans has no endpoint that revokes a token on its own, so PayMux
        expires the underlying transaction, which has the same effect.
      operationId: cancelCheckoutSession
      responses:
        "200":
          description: The payment after the session was retired.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Payment" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/v1/payments/{paymentId}/refunds:
    parameters:
      - $ref: "#/components/parameters/PaymentId"
    post:
      tags: [Refunds]
      summary: Refund a payment
      description: |
        Refunds part or all of a settled payment. Omit `amount` to refund the
        remaining balance. An `Idempotency-Key` doubles as the refund key, so a
        retried request cannot refund twice.
      operationId: createRefund
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateRefundRequest" }
      responses:
        "201":
          description: The refund.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Refund" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "422":
          description: The payment cannot be refunded, or the amount exceeds the balance.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }

    get:
      tags: [Refunds]
      summary: List a payment's refunds
      operationId: listRefunds
      responses:
        "200":
          description: The refunds.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/RefundList" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: http
      scheme: bearer
      description: |
        An application API key: `Authorization: Bearer pmx_live_…`.

        The key determines which application the request acts as. A `pmx_test_…`
        key works only against a sandbox gateway account, and a `pmx_live_…` key
        only against production.
  parameters:
    PaymentId:
      name: paymentId
      in: path
      required: true
      schema: { type: string, example: pay_01ARZ3NDEKTSV4RRFFQ69G5FAV }
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      description: |
        Makes the request safe to retry. The same key with the same body
        returns the original result; the same key with a different body is
        rejected as a conflict. Keys are remembered for 24 hours.
      schema: { type: string, maxLength: 255 }
    Limit:
      name: limit
      in: query
      description: Page size, 1–100. Defaults to 25.
      schema: { type: integer, minimum: 1, maximum: 100, default: 25 }
    StartingAfter:
      name: starting_after
      in: query
      description: Return the page following this identifier.
      schema: { type: string }
  responses:
    Unauthorized:
      description: The API key is missing, invalid, revoked or expired.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    NotFound:
      description: |
        No such resource. Also returned when the resource belongs to another
        application, so identifiers cannot be probed across tenants.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Conflict:
      description: The request conflicts with existing state.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    ValidationFailed:
      description: The request is well formed but contains invalid values.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
    GatewayNotConfigured:
      description: No gateway account is configured for this application.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              description: A stable, machine-readable identifier.
              examples: [PAYMENT_NOT_FOUND, VALIDATION_FAILED, RATE_LIMITED]
            message:
              type: string
              description: A human-readable explanation, safe to display.
            fields:
              type: array
              description: Which fields were invalid, when applicable.
              items:
                type: object
                properties:
                  field: { type: string }
                  message: { type: string }
            request_id:
              type: string
              description: Quote this when reporting a problem.
              example: req_01ARZ3NDEKTSV4RRFFQ69G5FAV
    PaymentStatus:
      type: string
      description: PayMux's normalized payment status.
      enum:
        - PENDING
        - AUTHORIZED
        - PAID
        - FAILED
        - CANCELED
        - EXPIRED
        - REFUNDED
        - PARTIALLY_REFUNDED
...
    Customer:
      type: object
      properties:
        first_name: { type: string }
        last_name: { type: string }
        email: { type: string, format: email }
        phone: { type: string }
        billing_address: { $ref: "#/components/schemas/Address" }
        shipping_address: { $ref: "#/components/schemas/Address" }
    Address:
      type: object
      properties:
        first_name: { type: string }
        last_name: { type: string }
        phone: { type: string }
        address: { type: string }
        city: { type: string }
        postal_code: { type: string }
        country_code: { type: string }
    Item:
      type: object
      required: [name, price, quantity]
      properties:
        id: { type: string, description: Your own SKU. }
        name: { type: string, maxLength: 50 }
        price:
          type: integer
          format: int64
          description: Unit price in the currency's minor unit.
        quantity: { type: integer, minimum: 1 }
        category: { type: string }
        merchant_name: { type: string }
        url: { type: string }
    GatewayOptions:
      type: object
      description: |
        Gateway-specific parameters, namespaced by gateway. Fields are
        validated; unknown fields are rejected rather than forwarded.
      properties:
        midtrans:
          type: object
          properties:
            credit_card:
              type: object
              properties:
                secure:
                  type: boolean
                  default: true
                  description: 3-D Secure. On unless deliberately disabled.
                bank: { type: string }
                channel: { type: string }
                save_card: { type: boolean }
            bank_transfer:
              type: object
              properties:
                bank:
                  type: string
                  enum: [bca, bni, bri, permata, cimb, mandiri, danamon, maybank, mega]
                va_number: { type: string }
                recipient_name: { type: string }
            gopay:
              type: object
              properties:
                enable_callback: { type: boolean }
                callback_url: { type: string }
            shopeepay:
              type: object
              properties:
                callback_url: { type: string }
            cstore:
              type: object
              properties:
                store: { type: string, enum: [indomaret, alfamart] }
                message: { type: string }
            page_expiry_minutes: { type: integer, minimum: 0 }
            error_redirect_url: { type: string }
    CreatePaymentRequest:
      type: object
      required: [application_order_id, amount]
      properties:
        application_order_id:
          type: string
          maxLength: 128
          description: |
            Your own order reference. Unique within your application only —
            two applications may both use `INV-001`.
        amount:
          type: integer
          format: int64
          minimum: 1
          description: Total in the currency's minor unit.
        currency: { type: string, default: IDR }
        customer: { $ref: "#/components/schemas/Customer" }
        items:
          type: array
          description: Line items. If present, they must sum to `amount`.
          items: { $ref: "#/components/schemas/Item" }
        enabled_payment_methods:
          type: array
          description: |
            Restrict the channels offered. Omit to use whatever the merchant
            account has activated.
          items: { type: string }
          example: [gopay, bca_va, bni_va]
        expires_at: { type: string, format: date-time }
        callback_url:
          type: string
          description: Where the gateway returns the payer after checkout.
        custom_fields:
          type: array
          maxItems: 3
          items: { type: string }
        metadata:
          type: object
          description: Your own data, stored and echoed back in events.
          additionalProperties: true
        gateway: { type: string, default: midtrans }
        gateway_options: { $ref: "#/components/schemas/GatewayOptions" }
    Payment:
      type: object
      properties:
        id: { type: string, example: pay_01ARZ3NDEKTSV4RRFFQ69G5FAV }
        object: { type: string, const: payment }
        application_id: { type: string }
        gateway: { type: string, example: midtrans }
        application_order_id: { type: string }
        gateway_order_id:
          type: string
          description: PayMux's globally unique order id, sent to the gateway.
        gateway_transaction_id: { type: string }
        amount: { type: integer, format: int64 }
        currency: { type: string }
        status: { $ref: "#/components/schemas/PaymentStatus" }
        gateway_status:
          type: string
          description: The gateway's own status, kept alongside the normalized one.
          example: settlement
        fraud_status: { type: string }
        payment_type: { type: string }
        payment_method: { type: string }
        snap_token: { type: string }
        redirect_url: { type: string }
        refunded_amount: { type: integer, format: int64 }
        refundable_amount: { type: integer, format: int64 }
        metadata: { type: object, additionalProperties: true }
        customer: { $ref: "#/components/schemas/Customer" }
        items:
          type: array
          items: { $ref: "#/components/schemas/Item" }
        expires_at: { type: [string, "null"], format: date-time }
        paid_at: { type: string, format: date-time }
        canceled_at: { type: string, format: date-time }
        expired_at: { type: string, format: date-time }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    CreateRefundRequest:
      type: object
      properties:
        amount:
          type: integer
          format: int64
          description: Minor units. Omit or send 0 to refund the remaining balance.
        reason: { type: string, maxLength: 255 }
    Refund:
      type: object
      properties:
        id: { type: string, example: rfd_01ARZ3NDEKTSV4RRFFQ69G5FAV }
        object: { type: string, const: refund }
        payment_id: { type: string }
        gateway_refund_id: { type: string }
        amount: { type: integer, format: int64 }
        currency: { type: string }
        reason: { type: string }
        status: { type: string, enum: [PENDING, SUCCEEDED, FAILED] }
        gateway_status: { type: string }
        failure_reason: { type: string }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
...
    ListEnvelope:
      type: object
      properties:
        has_more:
          type: boolean
          description: Whether a further page exists.
        limit: { type: integer }
    PaymentList:
      allOf:
        - $ref: "#/components/schemas/ListEnvelope"
        - type: object
          properties:
            data:
              type: array
              items: { $ref: "#/components/schemas/Payment" }
    RefundList:
      allOf:
        - $ref: "#/components/schemas/ListEnvelope"
        - type: object
          properties:
            data:
              type: array
              items: { $ref: "#/components/schemas/Refund" }
...